device/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0
pydantic-settings>=2.0
aiofiles>=23.0
python-multipart>=0.0.6
pyserial>=3.5
pyserial-asyncio>=0.6
httpx>=0.24.0
loguru>=0.7.0
sounddevice>=0.4.6
soundfile>=0.12.0
numpy>=1.24.0
